@@ -29,34 +29,57 @@ func TestQuery(t *testing.T) {
 		{
 			input: "../../tests/state/query/q1.json",
 			query: Query{
-				Filters: nil,
-				Sort:    nil,
-				Page:    Pagination{Limit: 2, Token: ""},
-				Filter:  nil,
+				QueryFields: QueryFields{
+					Filters: nil,
+					Sort:    nil,
+					Page:    Pagination{Limit: 2, Token: ""},
+				},
+				Filter: nil,
 			},
 		},
 		{
 			input: "../../tests/state/query/q2.json",
 			query: Query{
-				Filters: nil,
-				Sort:    nil,
-				Page:    Pagination{Limit: 2, Token: ""},
-				Filter:  &EQ{Key: "state", Val: "CA"},
+				QueryFields: QueryFields{
+					Filters: map[string]any{
+						"EQ": map[string]any{
+							"state": "CA",
+						},
+					},
+					Sort: nil,
+					Page: Pagination{Limit: 2, Token: ""},
+				},
+				Filter: &EQ{Key: "state", Val: "CA"},
 			},
 		},
 		{
 			input: "../../tests/state/query/q3.json",
 			query: Query{
-				Filters: nil,
-				Sort: []Sorting{
-					{Key: "state", Order: "DESC"},
-					{Key: "person.name", Order: ""},
+				QueryFields: QueryFields{
+					Filters: map[string]any{
+						"AND": []any{
+							map[string]any{
+								"EQ": map[string]any{
+									"person.org": "A",
+								},
+							},
+							map[string]any{
+								"IN": map[string]any{
+									"state": []any{"CA", "WA"},
+								},
+							},
+						},
+					},
+					Sort: []Sorting{
+						{Key: "state", Order: "DESC"},
+						{Key: "person.name", Order: ""},
+					},
+					Page: Pagination{Limit: 0, Token: ""},
 				},
-				Page: Pagination{Limit: 0, Token: ""},
 				Filter: &AND{
 					Filters: []Filter{
 						&EQ{Key: "person.org", Val: "A"},
-						&IN{Key: "state", Vals: []interface{}{"CA", "WA"}},
+						&IN{Key: "state", Vals: []any{"CA", "WA"}},
 					},
 				},
 			},
@@ -64,12 +87,36 @@ func TestQuery(t *testing.T) {
 		{
 			input: "../../tests/state/query/q4.json",
 			query: Query{
-				Filters: nil,
-				Sort: []Sorting{
-					{Key: "state", Order: "DESC"},
-					{Key: "person.name", Order: ""},
+				QueryFields: QueryFields{
+					Filters: map[string]any{
+						"OR": []any{
+							map[string]any{
+								"EQ": map[string]any{
+									"person.org": "A",
+								},
+							},
+							map[string]any{
+								"AND": []any{
+									map[string]any{
+										"EQ": map[string]any{
+											"person.org": "B",
+										},
+									},
+									map[string]any{
+										"IN": map[string]any{
+											"state": []any{"CA", "WA"},
+										},
+									},
+								},
+							},
+						},
+					},
+					Sort: []Sorting{
+						{Key: "state", Order: "DESC"},
+						{Key: "person.name", Order: ""},
+					},
+					Page: Pagination{Limit: 2, Token: ""},
 				},
-				Page: Pagination{Limit: 2, Token: ""},
 				Filter: &OR{
 					Filters: []Filter{
 						&EQ{Key: "person.org", Val: "A"},
