@@ -108,6 +108,9 @@ func TestGet(t *testing.T) {
 							"value": {
 								S: aws.String("some value"),
 							},
+							"etag": {
+								S: aws.String("1bdead4badc0ffee"),
+							},
 						},
 					}, nil
 				},
@@ -123,6 +126,7 @@ func TestGet(t *testing.T) {
 		out, err := ss.Get(req)
 		assert.Nil(t, err)
 		assert.Equal(t, []byte("some value"), out.Data)
+		assert.Equal(t, "1bdead4badc0ffee", *out.ETag)
 	})
 	t.Run("Successfully retrieve item (with unexpired ttl)", func(t *testing.T) {
 		ss := StateStore{
@@ -138,6 +142,9 @@ func TestGet(t *testing.T) {
 							},
 							"testAttributeName": {
 								N: aws.String("4074862051"),
+							},
+							"etag": {
+								S: aws.String("1bdead4badc0ffee"),
 							},
 						},
 					}, nil
@@ -155,6 +162,7 @@ func TestGet(t *testing.T) {
 		out, err := ss.Get(req)
 		assert.Nil(t, err)
 		assert.Equal(t, []byte("some value"), out.Data)
+		assert.Equal(t, "1bdead4badc0ffee", *out.ETag)
 	})
 	t.Run("Successfully retrieve item (with expired ttl)", func(t *testing.T) {
 		ss := StateStore{
@@ -170,6 +178,9 @@ func TestGet(t *testing.T) {
 							},
 							"testAttributeName": {
 								N: aws.String("35489251"),
+							},
+							"etag": {
+								S: aws.String("1bdead4badc0ffee"),
 							},
 						},
 					}, nil
@@ -187,6 +198,7 @@ func TestGet(t *testing.T) {
 		out, err := ss.Get(req)
 		assert.Nil(t, err)
 		assert.Nil(t, out.Data)
+		assert.Nil(t, out.ETag)
 	})
 	t.Run("Unsuccessfully get item", func(t *testing.T) {
 		ss := StateStore{
@@ -227,6 +239,7 @@ func TestGet(t *testing.T) {
 		out, err := ss.Get(req)
 		assert.Nil(t, err)
 		assert.Nil(t, out.Data)
+		assert.Nil(t, out.ETag)
 	})
 	t.Run("Unsuccessfully with no required key", func(t *testing.T) {
 		ss := StateStore{
@@ -264,15 +277,13 @@ func TestSet(t *testing.T) {
 		ss := StateStore{
 			client: &mockedDynamoDB{
 				PutItemFn: func(input *dynamodb.PutItemInput) (output *dynamodb.PutItemOutput, err error) {
-					assert.Equal(t, map[string]*dynamodb.AttributeValue{
-						"key": {
-							S: aws.String("key"),
-						},
-						"value": {
-							S: aws.String(`{"Value":"value"}`),
-						},
-					}, input.Item)
-					assert.Equal(t, len(input.Item), 2)
+					assert.Equal(t, dynamodb.AttributeValue{
+						S: aws.String("key"),
+					}, *input.Item["key"])
+					assert.Equal(t, dynamodb.AttributeValue{
+						S: aws.String(`{"Value":"value"}`),
+					}, *input.Item["value"])
+					assert.Equal(t, len(input.Item), 3)
 
 					return &dynamodb.PutItemOutput{
 						Attributes: map[string]*dynamodb.AttributeValue{
@@ -294,11 +305,161 @@ func TestSet(t *testing.T) {
 		assert.Nil(t, err)
 	})
 
+	t.Run("Successfully set item with matching etag", func(t *testing.T) {
+		ss := StateStore{
+			client: &mockedDynamoDB{
+				PutItemFn: func(input *dynamodb.PutItemInput) (output *dynamodb.PutItemOutput, err error) {
+					assert.Equal(t, dynamodb.AttributeValue{
+						S: aws.String("key"),
+					}, *input.Item["key"])
+					assert.Equal(t, dynamodb.AttributeValue{
+						S: aws.String(`{"Value":"value"}`),
+					}, *input.Item["value"])
+					assert.Equal(t, "etag = :etag", *input.ConditionExpression)
+					assert.Equal(t, &dynamodb.AttributeValue{
+						S: aws.String("1bdead4badc0ffee"),
+					}, input.ExpressionAttributeValues[":etag"])
+					assert.Equal(t, len(input.Item), 3)
+
+					return &dynamodb.PutItemOutput{
+						Attributes: map[string]*dynamodb.AttributeValue{
+							"key": {
+								S: aws.String("value"),
+							},
+						},
+					}, nil
+				},
+			},
+		}
+		etag := "1bdead4badc0ffee"
+		req := &state.SetRequest{
+			ETag: &etag,
+			Key:  "key",
+			Value: value{
+				Value: "value",
+			},
+		}
+		err := ss.Set(req)
+		assert.Nil(t, err)
+	})
+
+	t.Run("Unsuccessfully set item with mismatched etag", func(t *testing.T) {
+		ss := StateStore{
+			client: &mockedDynamoDB{
+				PutItemFn: func(input *dynamodb.PutItemInput) (output *dynamodb.PutItemOutput, err error) {
+					assert.Equal(t, dynamodb.AttributeValue{
+						S: aws.String("key"),
+					}, *input.Item["key"])
+					assert.Equal(t, dynamodb.AttributeValue{
+						S: aws.String(`{"Value":"value"}`),
+					}, *input.Item["value"])
+					assert.Equal(t, "etag = :etag", *input.ConditionExpression)
+					assert.Equal(t, &dynamodb.AttributeValue{
+						S: aws.String("bogusetag"),
+					}, input.ExpressionAttributeValues[":etag"])
+					assert.Equal(t, len(input.Item), 3)
+
+					var checkErr dynamodb.ConditionalCheckFailedException
+					return nil, &checkErr
+				},
+			},
+		}
+		etag := "bogusetag"
+		req := &state.SetRequest{
+			ETag: &etag,
+			Key:  "key",
+			Value: value{
+				Value: "value",
+			},
+		}
+
+		err := ss.Set(req)
+		assert.NotNil(t, err)
+		switch tagErr := err.(type) {
+		case *state.ETagError:
+			assert.Equal(t, tagErr.Kind(), state.ETagMismatch)
+		default:
+			assert.True(t, false)
+		}
+	})
+
+	t.Run("Successfully set item with first-write-concurrency", func(t *testing.T) {
+		ss := StateStore{
+			client: &mockedDynamoDB{
+				PutItemFn: func(input *dynamodb.PutItemInput) (output *dynamodb.PutItemOutput, err error) {
+					assert.Equal(t, dynamodb.AttributeValue{
+						S: aws.String("key"),
+					}, *input.Item["key"])
+					assert.Equal(t, dynamodb.AttributeValue{
+						S: aws.String(`{"Value":"value"}`),
+					}, *input.Item["value"])
+					assert.Equal(t, "attribute_not_exists(etag)", *input.ConditionExpression)
+					assert.Equal(t, len(input.Item), 3)
+
+					return &dynamodb.PutItemOutput{
+						Attributes: map[string]*dynamodb.AttributeValue{
+							"key": {
+								S: aws.String("value"),
+							},
+						},
+					}, nil
+				},
+			},
+		}
+		req := &state.SetRequest{
+			Key: "key",
+			Value: value{
+				Value: "value",
+			},
+			Options: state.SetStateOption{
+				Concurrency: state.FirstWrite,
+			},
+		}
+		err := ss.Set(req)
+		assert.Nil(t, err)
+	})
+
+	t.Run("Unsuccessfully set item with first-write-concurrency", func(t *testing.T) {
+		ss := StateStore{
+			client: &mockedDynamoDB{
+				PutItemFn: func(input *dynamodb.PutItemInput) (output *dynamodb.PutItemOutput, err error) {
+					assert.Equal(t, dynamodb.AttributeValue{
+						S: aws.String("key"),
+					}, *input.Item["key"])
+					assert.Equal(t, dynamodb.AttributeValue{
+						S: aws.String(`{"Value":"value"}`),
+					}, *input.Item["value"])
+					assert.Equal(t, "attribute_not_exists(etag)", *input.ConditionExpression)
+					assert.Equal(t, len(input.Item), 3)
+
+					var checkErr dynamodb.ConditionalCheckFailedException
+					return nil, &checkErr
+				},
+			},
+		}
+		req := &state.SetRequest{
+			Key: "key",
+			Value: value{
+				Value: "value",
+			},
+			Options: state.SetStateOption{
+				Concurrency: state.FirstWrite,
+			},
+		}
+		err := ss.Set(req)
+		assert.NotNil(t, err)
+		switch err.(type) {
+		case *state.ETagError:
+			assert.True(t, false)
+		default:
+		}
+	})
+
 	t.Run("Successfully set item with ttl = -1", func(t *testing.T) {
 		ss := StateStore{
 			client: &mockedDynamoDB{
 				PutItemFn: func(input *dynamodb.PutItemInput) (output *dynamodb.PutItemOutput, err error) {
-					assert.Equal(t, len(input.Item), 3)
+					assert.Equal(t, len(input.Item), 4)
 					result := DynamoDBItem{}
 					dynamodbattribute.UnmarshalMap(input.Item, &result)
 					assert.Equal(t, result.Key, "someKey")
@@ -333,7 +494,7 @@ func TestSet(t *testing.T) {
 		ss := StateStore{
 			client: &mockedDynamoDB{
 				PutItemFn: func(input *dynamodb.PutItemInput) (output *dynamodb.PutItemOutput, err error) {
-					assert.Equal(t, len(input.Item), 3)
+					assert.Equal(t, len(input.Item), 4)
 					result := DynamoDBItem{}
 					dynamodbattribute.UnmarshalMap(input.Item, &result)
 					assert.Equal(t, result.Key, "someKey")
@@ -386,15 +547,13 @@ func TestSet(t *testing.T) {
 		ss := StateStore{
 			client: &mockedDynamoDB{
 				PutItemFn: func(input *dynamodb.PutItemInput) (output *dynamodb.PutItemOutput, err error) {
-					assert.Equal(t, map[string]*dynamodb.AttributeValue{
-						"key": {
-							S: aws.String("someKey"),
-						},
-						"value": {
-							S: aws.String(`{"Value":"someValue"}`),
-						},
-					}, input.Item)
-					assert.Equal(t, len(input.Item), 2)
+					assert.Equal(t, dynamodb.AttributeValue{
+						S: aws.String("someKey"),
+					}, *input.Item["key"])
+					assert.Equal(t, dynamodb.AttributeValue{
+						S: aws.String(`{"Value":"someValue"}`),
+					}, *input.Item["value"])
+					assert.Equal(t, len(input.Item), 3)
 
 					return &dynamodb.PutItemOutput{
 						Attributes: map[string]*dynamodb.AttributeValue{
@@ -497,7 +656,16 @@ func TestBulkSet(t *testing.T) {
 							},
 						},
 					}
-					assert.Equal(t, expected, input.RequestItems)
+
+					for tbl := range expected {
+						for reqNum := range expected[tbl] {
+							expectedItem := expected[tbl][reqNum].PutRequest.Item
+							inputItem := input.RequestItems[tbl][reqNum].PutRequest.Item
+
+							assert.Equal(t, expectedItem["key"], inputItem["key"])
+							assert.Equal(t, expectedItem["value"], inputItem["value"])
+						}
+					}
 
 					return &dynamodb.BatchWriteItemOutput{
 						UnprocessedItems: map[string][]*dynamodb.WriteRequest{},
@@ -558,7 +726,15 @@ func TestBulkSet(t *testing.T) {
 							},
 						},
 					}
-					assert.Equal(t, expected, input.RequestItems)
+					for tbl := range expected {
+						for reqNum := range expected[tbl] {
+							expectedItem := expected[tbl][reqNum].PutRequest.Item
+							inputItem := input.RequestItems[tbl][reqNum].PutRequest.Item
+
+							assert.Equal(t, expectedItem["key"], inputItem["key"])
+							assert.Equal(t, expectedItem["value"], inputItem["value"])
+						}
+					}
 
 					return &dynamodb.BatchWriteItemOutput{
 						UnprocessedItems: map[string][]*dynamodb.WriteRequest{},
@@ -625,7 +801,15 @@ func TestBulkSet(t *testing.T) {
 							},
 						},
 					}
-					assert.Equal(t, expected, input.RequestItems)
+					for tbl := range expected {
+						for reqNum := range expected[tbl] {
+							expectedItem := expected[tbl][reqNum].PutRequest.Item
+							inputItem := input.RequestItems[tbl][reqNum].PutRequest.Item
+
+							assert.Equal(t, expectedItem["key"], inputItem["key"])
+							assert.Equal(t, expectedItem["value"], inputItem["value"])
+						}
+					}
 
 					return &dynamodb.BatchWriteItemOutput{
 						UnprocessedItems: map[string][]*dynamodb.WriteRequest{},
@@ -670,6 +854,12 @@ func TestBulkSet(t *testing.T) {
 					Value: "value",
 				},
 			},
+			{
+				Key: "key",
+				Value: value{
+					Value: "value",
+				},
+			},
 		}
 		err := ss.BulkSet(req)
 		assert.NotNil(t, err)
@@ -697,6 +887,69 @@ func TestDelete(t *testing.T) {
 		}
 		err := ss.Delete(req)
 		assert.Nil(t, err)
+	})
+
+	t.Run("Successfully delete item with matching etag", func(t *testing.T) {
+		etag := "1bdead4badc0ffee"
+		req := &state.DeleteRequest{
+			ETag: &etag,
+			Key:  "key",
+		}
+
+		ss := StateStore{
+			client: &mockedDynamoDB{
+				DeleteItemFn: func(input *dynamodb.DeleteItemInput) (output *dynamodb.DeleteItemOutput, err error) {
+					assert.Equal(t, map[string]*dynamodb.AttributeValue{
+						"key": {
+							S: aws.String(req.Key),
+						},
+					}, input.Key)
+					assert.Equal(t, "etag = :etag", *input.ConditionExpression)
+					assert.Equal(t, &dynamodb.AttributeValue{
+						S: aws.String("1bdead4badc0ffee"),
+					}, input.ExpressionAttributeValues[":etag"])
+
+					return nil, nil
+				},
+			},
+		}
+		err := ss.Delete(req)
+		assert.Nil(t, err)
+	})
+
+	t.Run("Unsuccessfully delete item with mismatched etag", func(t *testing.T) {
+		etag := "bogusetag"
+		req := &state.DeleteRequest{
+			ETag: &etag,
+			Key:  "key",
+		}
+
+		ss := StateStore{
+			client: &mockedDynamoDB{
+				DeleteItemFn: func(input *dynamodb.DeleteItemInput) (output *dynamodb.DeleteItemOutput, err error) {
+					assert.Equal(t, map[string]*dynamodb.AttributeValue{
+						"key": {
+							S: aws.String(req.Key),
+						},
+					}, input.Key)
+					assert.Equal(t, "etag = :etag", *input.ConditionExpression)
+					assert.Equal(t, &dynamodb.AttributeValue{
+						S: aws.String("bogusetag"),
+					}, input.ExpressionAttributeValues[":etag"])
+
+					var checkErr dynamodb.ConditionalCheckFailedException
+					return nil, &checkErr
+				},
+			},
+		}
+		err := ss.Delete(req)
+		assert.NotNil(t, err)
+		switch tagErr := err.(type) {
+		case *state.ETagError:
+			assert.Equal(t, tagErr.Kind(), state.ETagMismatch)
+		default:
+			assert.True(t, false)
+		}
 	})
 
 	t.Run("Unsuccessfully delete item", func(t *testing.T) {
@@ -771,6 +1024,9 @@ func TestBulkDelete(t *testing.T) {
 			},
 		}
 		req := []state.DeleteRequest{
+			{
+				Key: "key",
+			},
 			{
 				Key: "key",
 			},
