@@ -312,7 +312,7 @@ func TestServiceBusQueueMetadata(t *testing.T) {
 
 		// Send events that the application above will observe.
 		ctx.Log("Invoking binding!")
-		req := &daprClient.InvokeBindingRequest{Name: "sb-binding-1", Operation: "create", Data: []byte("test msg"), Metadata: map[string]string{"TestMetadata": "Some Metadata"}}
+		req := &daprClient.InvokeBindingRequest{Name: "sb-binding-1", Operation: "create", Data: []byte("test msg"), Metadata: map[string]string{"Testmetadata": "Some Metadata"}}
 		err = client.InvokeOutputBinding(ctx, req)
 		require.NoError(ctx, err, "error publishing message")
 
@@ -330,8 +330,8 @@ func TestServiceBusQueueMetadata(t *testing.T) {
 				messages.Observe(string(in.Data))
 				ctx.Logf("Got message: %s - %+v", string(in.Data), in.Metadata)
 				require.NotEmpty(t, in.Metadata)
-				require.Contains(t, in.Metadata, "TestMetadata")
-				require.Equal(t, "Some Metadata", in.Metadata["TestMetadata"])
+				require.Contains(t, in.Metadata, "Testmetadata")
+				require.Equal(t, "Some Metadata", in.Metadata["Testmetadata"])
 
 				return []byte("{}"), nil
 			}))
@@ -353,6 +353,36 @@ func TestServiceBusQueueMetadata(t *testing.T) {
 		Run()
 }
 
+func TestServiceBusQueueDisableEntityManagement(t *testing.T) {
+	ports, _ := dapr_testing.GetFreePorts(2)
+	grpcPort := ports[0]
+	httpPort := ports[1]
+
+	testWithExpectedFailure := func(ctx flow.Context) error {
+		client, err := daprClient.NewClientWithPort(fmt.Sprintf("%d", grpcPort))
+		require.NoError(t, err, "Could not initialize dapr client.")
+
+		// Send events that the application above will observe.
+		ctx.Log("Invoking binding!")
+		req := &daprClient.InvokeBindingRequest{Name: "mgmt-binding", Operation: "create", Data: []byte("test msg"), Metadata: map[string]string{"TestMetadata": "Some Metadata"}}
+		err = client.InvokeOutputBinding(ctx, req)
+		require.Error(ctx, err, "error publishing message")
+		return nil
+	}
+
+	flow.New(t, "servicebus queues certification - entity management disabled").
+		// Run the application logic above.
+		Step(sidecar.Run("metadataSidecar",
+			embedded.WithoutApp(),
+			embedded.WithDaprGRPCPort(grpcPort),
+			embedded.WithDaprHTTPPort(httpPort),
+			embedded.WithComponentsPath("./components/disable_entity_mgmt"),
+			componentRuntimeOptions(),
+		)).
+		Step("send and wait", testWithExpectedFailure).
+		Run()
+}
+
 func componentRuntimeOptions() []runtime.Option {
 	log := logger.NewLogger("dapr.components")
 
@@ -360,10 +390,10 @@ func componentRuntimeOptions() []runtime.Option {
 	bindingsRegistry.Logger = log
 	bindingsRegistry.RegisterInputBinding(func(l logger.Logger) bindings.InputBinding {
 		return binding_asb.NewAzureServiceBusQueues(l)
-	}, "azure.eventhubs")
+	}, "azure.servicebusqueues")
 	bindingsRegistry.RegisterOutputBinding(func(l logger.Logger) bindings.OutputBinding {
 		return binding_asb.NewAzureServiceBusQueues(l)
-	}, "azure.eventhubs")
+	}, "azure.servicebusqueues")
 
 	secretstoreRegistry := secretstores_loader.NewRegistry()
 	secretstoreRegistry.Logger = log
